Separate test for reverse_bits
